Day7-Backend

+++ dreamday/src/main/java/com/example/dreamday/Repository/UserRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.dreamday.Model.Users;
 
-public interface UserRepo extends JpaRepository<Users, Long> {
+public interface UserRepo extends JpaRepository<Users, Integer> {
     
 }
